requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asgi-lifespan==2.1.0
asttokens==3.0.0
asyncpg==0.30.0
autopep8==2.3.2
bcrypt==4.1.2
beautifulsoup4==4.12.3
black==25.1.0
cachetools==5.3.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coverage==7.7.1
cryptography==44.0.2
decorator==5.2.1
Deprecated==1.2.18
distro==1.9.0
docopt==0.6.2
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.12
fastapi-jwt-auth==0.5.0
flake8==7.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
iniconfig==2.1.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
jiter==0.9.0
limits==4.4.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mypy==1.15.0
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.69.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pipreqs==0.4.13
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.0
pydantic-settings==2.8.1
pydantic_core==2.33.0
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==1.7.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==5.0.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.3.0
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
respx==0.22.0
rsa==4.9
sentry-sdk==2.0.1
six==1.17.0
slowapi==0.1.8
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.46.1
syncpg==1.1.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
yarg==0.1.10
